backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
httpx>=0.25.0
numpy>=1.24.0
scikit-learn>=1.3.0
email-validator>=2.0.0 